requirements.txt
pydantic>=2.0,<3.0
anthropic>=0.40.0
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0
pytest>=8.0
pytest-asyncio>=0.25.0
